Move pipeline_config.py up a level.
