+++ src/invoke_training/scripts/invoke_train.py
@@ -4,7 +4,7 @@ from pathlib import Path
 import yaml
 from pydantic import TypeAdapter
 
-from invoke_training.config.pipelines.pipeline_config import PipelineConfig
+from invoke_training.config.pipeline_config import PipelineConfig
 from invoke_training.pipelines._experimental.sd_dpo_lora.train import (
     run_training as run_diffusion_dpo_sd,
 )
